+++ lib/features/developer/sign_up/ui/optional_data/developer_sign_up_optional_data_screen.dart
@@ -1,16 +1,16 @@
 import 'package:carrerk/core/helpers/extensions.dart';
 import 'package:carrerk/core/helpers/spacing.dart';
-import 'package:carrerk/core/routing/routes.dart';
 import 'package:carrerk/core/theming/styles.dart';
 import 'package:carrerk/core/widgets/app_back_icon.dart';
-import 'package:carrerk/core/widgets/app_check_box_and_agree_text.dart';
 import 'package:carrerk/core/widgets/app_text_button.dart';
-import 'package:carrerk/features/developer/sign_up/compoulsory_data/widgets/developer_compulsory_data_form.dart';
+import 'package:carrerk/features/developer/sign_up/ui/optional_data/widgets/optional_data_form.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
-class DeveloperSignUpCompulsoryDataScreen extends StatelessWidget {
-  const DeveloperSignUpCompulsoryDataScreen({super.key});
+import '../../../../../core/routing/routes.dart';
+
+class DeveloperSignUpOptionalDataScreen extends StatelessWidget {
+  const DeveloperSignUpOptionalDataScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -19,31 +19,27 @@ class DeveloperSignUpCompulsoryDataScreen extends StatelessWidget {
           child: SingleChildScrollView(
         child: Padding(
           padding: EdgeInsets.symmetric(horizontal: 16.0.w, vertical: 16.h),
-          //TODO: Modify spaces when validation return message appear "Overflow"
           child: Column(
             children: [
               const AppBackIcon(),
-              verticalSpace(16),
               Text(
-                'Developer',
+                'Optional Info.',
                 style: AppTextStyles.font24DunePoppinsMedium,
               ),
               verticalSpace(16),
               Text(
-                'Please enter the following data',
+                'Press continue to skip',
                 style: AppTextStyles.font15LiverPoppinsMedium,
               ),
+              verticalSpace(16),
+              const OptionalDataForm(),
               verticalSpace(32),
-              const DeveloperCompulsoryDataForm(),
-              verticalSpace(32),
-              const AppCheckBoxAndAgreeText(),
-              verticalSpace(24),
               AppTextButton(
-                  buttonText: 'Next Step',
+                  buttonText: 'Continue',
                   textStyle: AppTextStyles.font14WhitePoppinsMedium,
                   onPressed: () {
                     //TODO: Check Validation that data is correct to go to next page
-                    context.pushNamed(Routes.developerSignUpFillProfileScreen);
+                    context.pushNamed(Routes.developerSignUpSelectedCourses);
                   })
             ],
           ),
